order totals endpoint working

+++ src/ApiHw/Models/OrderPrices.cs
@@ -8,7 +8,7 @@ namespace ApiHw.Models
     public class OrderPrices
     {
         public string OrderId { get; set; }
-        public double OrderTotal { get; set; }
+        public decimal OrderTotal { get; set; }
         public List<OrderItemPrices> OrderItemPrices { get; set; }
 
         public OrderPrices()
@@ -20,7 +20,7 @@ namespace ApiHw.Models
 
     public class OrderItemPrices
     {
-        public double Price { get; set; }
+        public decimal Price { get; set; }
 
         public string ItemName { get; set; }
     }
